requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.7
google-generativeai==0.7.2
faiss-cpu==8.8.0.post1
pypdf==4.3.1
python-multipart==0.0.6
streamlit==1.32.0
requests==2.32.5
